+++ src/main/resources/schema.sql
@@ -57,13 +57,104 @@ CREATE TABLE films_likes (
     PRIMARY KEY (film_id, user_id)
 );
 
-CREATE TABLE IF NOT EXISTS directors (
-    director_id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    name VARCHAR(255) NOT NULL
-);
+CREATE TABLE IF NOT EXISTS directors
+(
+    director_id
+    BIGINT
+    GENERATED
+    BY
+    DEFAULT AS
+    IDENTITY
+    PRIMARY
+    KEY,
+    name
+    VARCHAR
+(
+    255
+) NOT NULL
+    );
 
-CREATE TABLE IF NOT EXISTS films_directors (
-    film_id BIGINT REFERENCES films (film_id) ON DELETE CASCADE,
-    director_id BIGINT REFERENCES directors (director_id) ON DELETE CASCADE,
-    PRIMARY KEY(film_id, director_id)
-);
+CREATE TABLE IF NOT EXISTS films_directors
+(
+    film_id BIGINT REFERENCES films
+(
+    film_id
+) ON DELETE CASCADE,
+    director_id BIGINT REFERENCES directors
+(
+    director_id
+)
+  ON DELETE CASCADE,
+    PRIMARY KEY
+(
+    film_id,
+    director_id
+)
+    );
+
+CREATE TABLE IF NOT EXISTS reviews
+(
+    review_id
+    BIGINT
+    generated
+    by
+    default as
+    identity
+    primary
+    key,
+    content
+    VARCHAR,
+    is_positive
+    BOOLEAN,
+    user_id
+    BIGINT,
+    film_id
+    BIGINT,
+    useful
+    INTEGER,
+    FOREIGN
+    KEY
+(
+    film_id
+) REFERENCES films
+(
+    film_id
+),
+    FOREIGN KEY
+(
+    user_id
+) REFERENCES users
+(
+    user_id
+)
+    );
+
+CREATE TABLE IF NOT EXISTS users_reviews
+(
+    user_id
+    BIGINT,
+    review_id
+    BIGINT,
+    is_helpful
+    BOOLEAN,
+    PRIMARY
+    KEY
+(
+    user_id,
+    review_id
+),
+    FOREIGN KEY
+(
+    user_id
+) REFERENCES users
+(
+    user_id
+),
+    FOREIGN KEY
+(
+    review_id
+) REFERENCES reviews
+(
+    review_id
+)
+    );
